fix: use precise float frames for exact animation duration
Problem:
- GPU worker used int(frame_rate * duracao) which truncates decimals
- VPS worker used Math.round(image.duracao * fps) which rounds
- FFmpeg renders using exact duration in seconds (-t duracao)
- This caused animation mismatch: zoom calculated for N frames but video rendered for M frames

Example:
- Duration: 3.33s, FPS: 24
- Real frames: 79.92
- Old GPU: int(79.92) = 79 → animation completes early
- Old VPS: round(79.92) = 80 → animation doesn't complete
- FFmpeg renders: 79.92 frames (3.33s)

Solution:
- Use float for totalFrames: frame_rate * duracao (no rounding)
- Animation formulas now use precise value: on/79.92
- FFmpeg renders exact duration: -t 3.33
- Result: animation completes perfectly at video end

Changes:
- GPU (rp_handler.py): total_frames = frame_rate * duracao (float)
- VPS (localVideoProcessor.ts): totalFrames = image.duracao * fps (float)

Impact:
- Fixes 70% of durations that aren't exact multiples of 1/24
- Eliminates "jump" or premature completion on last frame
- Animation now matches video duration exactly

+++ src/worker-python/rp_handler.py
@@ -422,7 +422,8 @@ def image_to_video(
         image_metadata = get_image_metadata(image_path)
 
         # Zoom parameters - Optimized upscale (6x) for balanced quality and performance
-        total_frames = int(frame_rate * duracao)
+        # Use FLOAT for precise animation timing - no rounding to ensure animation completes exactly at video end
+        total_frames = frame_rate * duracao  # e.g., 24 * 3.33 = 79.92 frames (precise)
         upscale_factor = 6  # Balanced upscale: 6x for good quality and faster processing
 
         # Use actual image dimensions if available, otherwise default to 1920x1080
